// frontend/src/enums/review.enum.ts
export enum ReviewSearchStatus {
  ALL = 'ALL',
  ONE_STAR = 'ONE_STAR',
  TWO_STAR = 'TWO_STAR',
  THREE_STAR = 'THREE_STAR',
  FOUR_STAR = 'FOUR_STAR',
  FIVE_STAR = 'FIVE_STAR',
  WITH_IMAGES = 'WITH_IMAGES',
}

export enum ReplyStatus {
  PENDING = 'PENDING',
  REPLIED = 'REPLIED',
}

export enum DateFilter {
  TODAY = 'Today',
  YESTERDAY = 'Yesterday',
  LAST_7_DAYS = 'Last 7 days',
  LAST_30_DAYS = 'Last 30 days',
  THIS_MONTH = 'This month',
  LAST_MONTH = 'Last month',
  THIS_YEAR = 'This year',
  LAST_YEAR = 'Last year',
  ALL_TIME = 'All time',
}